+++ packages/core/src/types/agent.ts
@@ -2,5 +2,11 @@ export type Agent = {
   isAi: boolean;
   id: string | null;
   name: string;
-  avatar: string | null;
+  /** Avatar URL or a data URL */
+  avatarUrl: string | null;
+  /**
+   *  Avatar URL or a data URL
+   *  @deprecated Use `avatarUrl` instead
+   */
+  avatar?: string | null;
 };