+++ packages/react-headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opencx/widget-react-headless
 
+## 4.0.33
+
+### Patch Changes
+
+- introduce `bot.avatarUrl` and deprecate `bot.avatar`
+- Updated dependencies
+  - @opencx/widget-core@4.0.33
+
 ## 4.0.32
 
 ### Patch Changes